feat!: generates a UUID(v4) using built-in `randomUUID()` to track request
UUID is added to `req.uuid`. It is recommended that you use this for any error handling, so you can match up the request logged with the error. You can also simply use the same info here to capture in your error, and pass it to jsout in the 'context' property.

This is not technically a breaking change, but wanted to force a major version bump, just in case the very minor performance penalty or modifying `req` was an issue for anyone. I think it's important to make folks aware of the change, since Express does not garuntee any contract/type safety around `req`.