revert: reverted uuid feature
Realized it isn't very useful to have a uuid when it is the last middleware in the stack -- you can't really do anything with it other than reference it externally. Decided it isn't worth the performance penalty, however small, of generating a uuid for every request, even though we were using a native function. The feature was only out for a few minutes, so I'm not bumping the major version.